move common static functions for all Titan modules to Titan runtime
