+++ .gitignore
@@ -5,3 +5,4 @@
 # lua
 lua/src/lua
 lua/src/luac
+lua-5.3.4
