+++ .travis.yml
@@ -16,7 +16,9 @@ before_install:
   - pip install hererocks
   - hererocks lua_install -rlatest --$INSTALL_LUA
   - export PATH=$PATH:$PWD/lua_install/bin # Add directory with all installed binaries to PATH
-  - cd lua; make $TRAVIS_OS_NAME MYCFLAGS=-fPIC; cd ..
+  - wget http://www.lua.org/ftp/lua-5.3.4.tar.gz
+  - tar zxvf lua-5.3.4.tar.gz
+  - cd lua-5.3.4; make $TRAVIS_OS_NAME MYCFLAGS=-fPIC; cd ..
 
 install:
   - luarocks install busted
