+++ spec/coder_spec.lua
@@ -25,7 +25,7 @@ local function generate_modules(modules, main)
 end
 
 local function call(modname, code)
-    local cmd = string.format("lua/src/lua -l %s -e \"print(pcall(function () %s end))\"",
+    local cmd = string.format("lua-5.3.4/src/lua -l %s -e \"print(pcall(function () %s end))\"",
         modname, code)
     local f = io.popen(cmd)
     local out = f:read()
